chnages latest

+++ src/main/java/com/example/user/model/Customer.java
@@ -12,7 +12,7 @@ public class Customer {
 	private String name;
 	private String email;
 	private long phone;
-	private long credits;
+	private double credits;
 	
 	public Customer(String name, String email, long phone, long credits) {
 		super();
@@ -59,11 +59,11 @@ public class Customer {
 		this.phone = phone;
 	}
 
-	public long getCredits() {
+	public double getCredits() {
 		return credits;
 	}
 
-	public void setCredits(long credits) {
+	public void setCredits(double credits) {
 		this.credits = credits;
 	}
